WIP add middleware generation to templates
